backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.12
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg[binary]==3.2.3
pydantic==2.11.2
pydantic-settings==2.7.0
dramatiq[redis]==1.18.0
redis==5.2.1
openai==1.97.1
minio==7.2.12
pillow==11.1.0
PyMuPDF==1.25.2
svix==1.45.0
httpx==0.28.1
requests==2.32.3
pytest==8.3.4
pytest-asyncio==0.25.2
tenacity==9.0.0
numpy==2.2.2
clerk-backend-api==3.1.1
sentry-sdk[fastapi]==2.13.0
python-jose[cryptography]==3.3.0

# MCP server SDK for prompt_server (need server.fastmcp)
mcp>=1.2.0,<2.0.0

# Some environments default to psycopg2; include it to avoid import errors
psycopg2-binary==2.9.9
